Change plugin directory and fix documentation
While using `/usr/lib` might have been better in terms of UNIX
philosophy. It's not ideal as the default directory for plugins really
is in `/opt`. Let's use that instead and stay within the current
patterns followed by krakend. Even if they're not UNIXy.

This also updates the docs to reflect both this change and previous ones
related to assumptions for the image in terms of krakend's flexible
configuration.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>