Authz plugin work
This starts taking into use permissions-api in order build the
authorization plugin.

This requires a bunch of changes as we start looking deeper into krakend
integration...

We now use a Request/Response Modifier plugin instead of a
Server or Client plugin. This is because we will add the authz configuration
at the endpoint level and not globally nor at the backend level.

This then required some changes into how the API looks like and how
the test configuration is set.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>